Adds GA Accepted BR266 v.1.0 to default RM models (#252)
* refactors and refines menus for op files

* adds GA Accepted BR266 v.1.0.xml to default reference material models
