+++ squidApp/src/main/resources/org/cirdles/squid/gui/ProjectManager.fxml
@@ -26,8 +26,8 @@
       </HBox>
       <HBox spacing="10.0">
          <children>
-            <Label text="Prawn XML file:" />
-            <TextField fx:id="orignalPrawnFileName" prefHeight="27.0" prefWidth="775.0" promptText="name of selected Prawn XML file will appear here" HBox.hgrow="ALWAYS">
+            <Label text="Prawn source file:" />
+            <TextField fx:id="orignalPrawnFileName" prefHeight="27.0" prefWidth="775.0" promptText="name of selected Prawn XML or OP file will appear here" HBox.hgrow="ALWAYS">
                <font>
                   <Font name="Lucida Sans Regular" size="12.0" />
                </font>
