+++ squidApp/src/main/resources/org/cirdles/squid/gui/SquidUIController.fxml
@@ -45,7 +45,7 @@
               <items>
                   <MenuItem mnemonicParsing="false" onAction="#manageSpotsMenuItemAction" text="Manage Spots &amp; Reference Materials" />
                 <MenuItem mnemonicParsing="false" onAction="#auditSessionMenuItemAction" text="Manage Sample Naming" />
-                  <MenuItem mnemonicParsing="false" onAction="#auditMassesMenuItemAction" text="Audit Raw Data" />
+                  <MenuItem fx:id="auditRawDataMenuItem" mnemonicParsing="false" onAction="#auditMassesMenuItemAction" text="Audit Raw Data" />
                   <SeparatorMenuItem mnemonicParsing="false" />
                   <MenuItem fx:id="savePrawnFileCopyMenuItem" mnemonicParsing="false" onAction="#savePrawnFileCopyMenuItemAction" text="Save Project Prawn File as ..." />
                   <SeparatorMenuItem mnemonicParsing="false" />
