+++ squidCore/src/main/resources/org/cirdles/squid/parameters/parameterModels/referenceMaterialModels/GA Accepted BR266 v.1.0.xml
@@ -1,0 +1,63 @@
+<ReferenceMaterialModel>
+  <modelName>GA Accepted BR266</modelName>
+  <referenceDates>true</referenceDates>
+  <labName>Geoscience Australia</labName>
+  <version>1.0</version>
+  <dateCertified>2019-01-01</dateCertified>
+  <comments></comments>
+  <references></references>
+  <values>
+    <org.cirdles.squid.parameters.valueModels.Age206__238r>
+      <name>age206_238r</name>
+      <uncertaintyType>ABS</uncertaintyType>
+      <reference></reference>
+      <value>0</value>
+      <oneSigma>0</oneSigma>
+    </org.cirdles.squid.parameters.valueModels.Age206__238r>
+    <org.cirdles.squid.parameters.valueModels.Age207__206r>
+      <name>age207_206r</name>
+      <uncertaintyType>ABS</uncertaintyType>
+      <reference></reference>
+      <value>0</value>
+      <oneSigma>0</oneSigma>
+    </org.cirdles.squid.parameters.valueModels.Age207__206r>
+    <org.cirdles.squid.parameters.valueModels.Age207__235r>
+      <name>age207_235r</name>
+      <uncertaintyType>ABS</uncertaintyType>
+      <reference></reference>
+      <value>0</value>
+      <oneSigma>0</oneSigma>
+    </org.cirdles.squid.parameters.valueModels.Age207__235r>
+    <org.cirdles.squid.parameters.valueModels.Age208__232r>
+      <name>age208_232r</name>
+      <uncertaintyType>ABS</uncertaintyType>
+      <reference></reference>
+      <value>0</value>
+      <oneSigma>0</oneSigma>
+    </org.cirdles.squid.parameters.valueModels.Age208__232r>
+  </values>
+  <rhos/>
+  <isEditable>true</isEditable>
+  <concentrations>
+    <ValueModel>
+      <name>concU</name>
+      <uncertaintyType>ABS</uncertaintyType>
+      <reference></reference>
+      <value>903</value>
+      <oneSigma>0</oneSigma>
+    </ValueModel>
+    <ValueModel>
+      <name>concTh</name>
+      <uncertaintyType>ABS</uncertaintyType>
+      <reference></reference>
+      <value>0</value>
+      <oneSigma>0</oneSigma>
+    </ValueModel>
+  </concentrations>
+  <dataMeasured>
+    <boolean>false</boolean>
+    <boolean>false</boolean>
+    <boolean>false</boolean>
+    <boolean>false</boolean>
+  </dataMeasured>
+</ReferenceMaterialModel>
