+++ squidCore/src/main/resources/org/cirdles/squid/parameters/parameterModels/referenceMaterialModels/listOfSampleReferenceMaterials.txt
@@ -1,2 +1,3 @@
 Zircon-91500 v.1.0.xml
 Temora Placeholder v.1.0.xml
+GA Accepted BR266 v.1.0.xml
